refactor login.go for triage metadata PR
